+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Changes
 
+## Master
+
+* Add support for Load Balancer Label Selector targets
+* Add support for Load Balancer IP targets
+
 ## v1.19.0
 
 * Fix nil pointer dereference when creating a Load Balancer with HTTP(S)
@@ -7,8 +12,6 @@
 * Add `IncludedTraffic`, `OutgoingTraffic` and `IngoingTraffic` fields to `LoadBalancer`
 * Add `ChangeType()` method to the Load Balancer client
 * Fix retrying of requests that contain a body
-* Add support for Load Balancer Label Selector targets
-* Add support for Load Balancer IP Selector targets
 
 ## v1.18.2